"\n\nCommit: Refactor gait.py and test_gait"
